+++ src/main/kotlin/no/nav/amt/person/service/api/auth/AuthService.kt
@@ -1,4 +1,4 @@
-package no.nav.amt.person.service.controller.auth
+package no.nav.amt.person.service.api.auth
 
 import no.nav.security.token.support.core.context.TokenValidationContextHolder
 import no.nav.security.token.support.spring.validation.interceptor.JwtTokenUnauthorizedException
