+++ src/main/kotlin/no/nav/amt/person/service/api/dto/ArrangorAnsattDto.kt
@@ -1,4 +1,4 @@
-package no.nav.amt.person.service.controller.dto
+package no.nav.amt.person.service.api.dto
 
 import no.nav.amt.person.service.person.model.Person
 import java.util.UUID
