+++ src/main/kotlin/no/nav/amt/person/service/api/request/ArrangorAnsattRequest.kt
@@ -1,4 +1,4 @@
-package no.nav.amt.person.service.controller.request
+package no.nav.amt.person.service.api.request
 
 data class ArrangorAnsattRequest(
 	val personident: String,
